Fix process table snapshot module split
