+++ src/shared/child-process/__fixtures__/windows-console-visibility-allowlist.txt
@@ -63,7 +63,7 @@ relay/subprocess-tree-termination.ts
 relay/windows-port-scan.ts
 relay/workspace-space-scan.ts
 shared/fish-binary-requirement.ts
-shared/process-table-snapshot.ts
+shared/process-table-snapshot-reader.ts
 shared/pty-slave-line-discipline-echo.ts
 shared/ripgrep-process-availability.ts
 shared/shell-process-readiness.ts
